Inital commit: Includes User Signup and Validation
Initial setup of Spring Boot backend project to provide REST apis for the hoaxify app.
This commit is for User Signup and related validation.
Working in tandem on a React frontend which will be in a different repo.
This is an exercise to brush up on my Spring Boot and React skills.